requirements.txt
openai>=1.30.0
pydantic>=2.7.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.0.0
